Big Update to Solver & ReadMe
The solver algorithm has changed quite a bit. The get_anagram function
contains most of the changes.
It know looks for every anagram lower then the length of the conundrum.
Which means the time complexity has increased by a couple of seconds.
I'll try to decrease the time it takes in the next update.
The readme file has been updated with more info, resources etc...